+++ client/src/__tests__/components/services/SignUpAPI.test.js
@@ -5,7 +5,14 @@ jest.mock('axios');
 let signUpAPI;
 
 beforeEach(() => {
-  window.location.hostname = 'billdivider.fun';
+  delete window.location
+  Object.defineProperty(window, 'location', {
+    writable: true,
+    value: {
+      hostname: 'billdivider.fun'
+    },
+  });
+  
   signUpAPI = new SignUpAPI();
 });
 
@@ -15,11 +22,11 @@ afterEach(() => {
 });
 
 describe('signUp(user)', () => {
-  it('should call axios.post()', async () => {
+  it('should call axios.post() with the appropriate parameters', async () => {
     axios.post.mockImplementationOnce(() => { return { success: true, data: {} } });
     const newUser = { username: 'someUsername', password: 'somePassword' };
 
     await signUpAPI.signUp(newUser);
-    expect(axios.post).toHaveBeenCalledWith(`http://${window.location.hostname}:5000/api/users/signup`, newUser);
+    expect(axios.post).toHaveBeenCalledWith(`https://${window.location.hostname}/api/users/signup`, newUser);
   });
 });