+++ client/src/js/components/history/deleteSession.js
@@ -12,8 +12,7 @@ async function deleteSession(sessionID) {
   const loginToken = locateLoginToken();
 
   if(!loginToken) { // Not logged in - Redirecting...
-    window.location.href = 'signIn.html';
-    return ;
+    return redirectAfterDelayMillisecond('signIn.html', 1000, 'Not logged in');
   };
 
   try {
@@ -34,7 +33,7 @@ async function deleteSession(sessionID) {
       cookies.remove('loginToken');
       return redirectAfterDelayMillisecond('signIn.html', 1000, 'Not logged in');
 
-    } else { // Most likely 500
+    } else { // 500
       cookies.remove('loginToken');
       return redirectAfterDelayMillisecond('signIn.html');
     };